requirements.txt
Flask
tensorflow
numpy
Pillow
gunicorn
requests
boto3==1.28.38